split service in two
