+++ app/services/credits/applied_prepaid_credits_service.rb
@@ -2,11 +2,10 @@
 
 module Credits
   class AppliedPrepaidCreditsService < BaseService
-    Result = BaseResult[:prepaid_credit_amount_cents, :wallet_transactions, :wallet_transactions_preview]
+    Result = BaseResult[:prepaid_credit_amount_cents, :wallet_transactions]
 
-    def initialize(invoice:, preview: false)
+    def initialize(invoice:)
       @invoice = invoice
-      @preview = preview
 
       super(nil)
     end
@@ -18,19 +17,13 @@ module Credits
 
       result.prepaid_credit_amount_cents ||= 0
       result.wallet_transactions ||= []
-      result.wallet_transactions_preview = {}
 
       return result if wallets.empty?
-
-      if preview
-        result.wallet_transactions_preview = calculate_wallet_transactions
-        return result
-      end
 
       ActiveRecord::Base.transaction do
         Customers::LockService.call(customer:, scope: :prepaid_credit) do
           # per each wallet we create a single wallet transaction. wallets_transactions is a hash with wallet and total transaction amount
-          wallets_transactions = calculate_wallet_transactions
+          wallets_transactions = CalculateApplicableWalletTransactionsService.call!(invoice:).wallet_transactions
 
           wallets_transactions.each do |wallet, amount_cents|
             wallet_transaction = create_wallet_transaction(wallet, amount_cents)
@@ -56,50 +49,9 @@ module Credits
 
     private
 
-    attr_accessor :invoice, :preview
+    attr_accessor :invoice
 
     delegate :customer, to: :invoice
-
-    def calculate_wallet_transactions
-      ordered_remaining_amounts = calculate_amounts_for_fees_by_type_and_bm
-      remaining_invoice_amount = invoice.total_amount_cents
-      wallets_transactions = {}
-
-      wallets.each do |wallet|
-        wallet.reload
-        wallet_fee_transactions = []
-        wallet_targets_array = wallet.wallet_targets.map do |wt|
-          if wt&.billable_metric_id
-            ["charge", wt.billable_metric_id]
-          end
-        end
-        wallet_types_array = wallet.allowed_fee_types
-
-        ordered_remaining_amounts.each do |fee_key, remaining_amount|
-          next if remaining_amount <= 0
-
-          next unless applicable_fee?(fee_key:, targets: wallet_targets_array, types: wallet_types_array, wallet:)
-
-          used_amount = wallet_fee_transactions.sum { |t| t[:amount_cents] }
-          remaining_wallet_balance = wallet.balance_cents - used_amount
-          next if remaining_wallet_balance <= 0
-
-          transaction_amount = [remaining_amount, remaining_wallet_balance, remaining_invoice_amount].min
-          next if transaction_amount <= 0
-
-          ordered_remaining_amounts[fee_key] -= transaction_amount
-          remaining_invoice_amount -= transaction_amount
-          wallet_fee_transactions << {
-            fee_key: fee_key,
-            amount_cents: transaction_amount
-          }
-        end
-        total_amount_cents = wallet_fee_transactions.sum { |t| t[:amount_cents] }
-        next if total_amount_cents <= 0
-        wallets_transactions[wallet] = total_amount_cents
-      end
-      wallets_transactions
-    end
 
     def schedule_webhook_notifications(wallet_transactions)
       wallet_transactions.each do |wt|
@@ -140,51 +92,6 @@ module Credits
       invoice.prepaid_purchased_credit_amount_cents = purchased_amount if purchased_amount > 0
     end
 
-    def calculate_amounts_for_fees_by_type_and_bm
-      remaining = Hash.new(0)
-      fees = invoice.persisted? ? invoice.fees.includes(:charge) : invoice.fees
-
-      fees.each do |fee|
-        next if fee.sub_total_excluding_taxes_amount_cents == 0
-
-        cap = fee.sub_total_excluding_taxes_amount_cents +
-          fee.taxes_precise_amount_cents -
-          fee.precise_credit_notes_amount_cents
-
-        next if cap <= 0
-        key = [fee.fee_type, fee.charge&.billable_metric_id]
-        if fee.organization.events_targeting_wallets_enabled? && fee.charge&.accepts_target_wallet
-          key << fee.grouped_by&.dig("target_wallet_code")
-        end
-        remaining[key] += cap
-      end
-
-      ordered = remaining.sort_by { |_, v| -v }.to_h
-      reconcile_remaining_amounts(ordered)
-    end
-
-    def reconcile_remaining_amounts(ordered_remaining_amounts)
-      return ordered_remaining_amounts if ordered_remaining_amounts.empty?
-
-      precise_total = ordered_remaining_amounts.values.sum
-      difference = invoice.total_amount_cents - precise_total
-
-      # Only reconcile small rounding differences (at most 1 cent per fee bucket).
-      return ordered_remaining_amounts if difference <= 0
-      return ordered_remaining_amounts if difference > ordered_remaining_amounts.size
-
-      largest_key = ordered_remaining_amounts.keys.first
-      ordered_remaining_amounts[largest_key] += difference
-      ordered_remaining_amounts
-    end
-
-    def wallets_already_applied?
-      return false unless invoice
-      return false unless invoice.persisted?
-
-      WalletTransaction.exists?(invoice_id: invoice.id, wallet_id: wallets.map(&:id))
-    end
-
     def create_wallet_transaction(wallet, amount_cents)
       wallet_credit = WalletCredit.from_amount_cents(wallet:, amount_cents:)
 
@@ -200,28 +107,19 @@ module Credits
       result.wallet_transaction
     end
 
-    def applicable_fee?(fee_key:, targets:, types:, wallet:)
-      target_wallet_code = fee_key[2]
-
-      # If fee has target_wallet_code, only matching wallet can apply credits
-      if target_wallet_code.present?
-        return wallet.code == target_wallet_code
-      end
-
-      fee_key_without_wallet = fee_key.first(2)
-      target_match = targets.include?(fee_key_without_wallet)
-      type_match = types.include?(fee_key.first)
-      unrestricted_wallet = targets.empty? && types.empty?
-
-      target_match || type_match || unrestricted_wallet
-    end
-
     def wallets
       @wallets ||= begin
         scope = customer.wallets.active.includes(:wallet_targets).with_positive_balance
         scope = scope.where(balance_currency: invoice.currency) if invoice.currency.present?
         scope.in_application_order
       end
+    end
+
+    def wallets_already_applied?
+      return false unless invoice
+      return false unless invoice.persisted?
+
+      WalletTransaction.exists?(invoice_id: invoice.id, wallet_id: wallets.map(&:id))
     end
   end
 end
